Crate rust_wasm_websocket

rust_wasm_websocket

crate for wasm websocket
repo; version: 0.5.2 date: 2021-01-15 authors: Luciano Bestia

websocket

The library contains the generic parts.
The specific implementation needs to implement the WebSocketTrait.

cargo crev reviews and advisory

It is recommended to always use cargo-crev
to verify the trustworthiness of each of your dependencies.
Please, spread this info.
On the web use this url to read crate reviews. Example:
https://web.crev.dev/rust-reviews/crate/num-traits/

Modules

websocket_mod
module that cares about WebSocket communication